// types/app-state.ts
export interface AppState {
  showSideMenu: boolean;
  setShowSideMenu: (value: boolean) => void;
  showPopup: boolean;
  setShowPopup: (value: boolean) => void;
  inspectionId: string;
  setInspectionId: (value: string) => void;
  inspectionData: InspectionData[];
  setInspectionData: (value: InspectionData[]) => void;
  popupType: PopupTypes | undefined;
  setPopupType: (value: PopupTypes) => void;
  storedQueue: VideoQueueData[];
  setStoredQueue: (queue: VideoQueueData[]) => void;
  showProgressStatus: { full: boolean; half: boolean };
  setShowProgressStatus: (value: { full: boolean; half: boolean }) => void;
  imageUrl: string;
  setImageUrl: (value: string) => void;
  userId: string;
  setUserId: (value: string) => void;
  organization: OrganizationInfo | undefined;
  setOrganization: (value: OrganizationInfo) => void;
  allowedActions: AllowedActions;
  setAllowedActions: (value: AllowedActions) => void;
  isLogoutInitiated: boolean;
  setIsLogoutInitiated: (value: boolean) => void;
  showOffCanvas: boolean;
  setShowOffCanvas: (value: boolean) => void;
  offCanvasType: OffCanvasTypes | undefined;
  setOffCanvasType: (value: OffCanvasTypes) => void;
  customers: Customer[];
  setCustomers: (value: Customer[]) => void;
  isCustomerUser: boolean;
  setIsCustomerUser: (value: boolean) => void;
  isCustomerView: boolean;
  setIsCustomerView: (value: boolean) => void;
  userRole: UserRole | undefined;
  setUserRole: (value: UserRole) => void;
  customerId: string | undefined;
  setCustomerId: (value: string) => void;
  customerName: string | undefined;
  setCustomerName: (value: string | undefined) => void;
  isUserAdded: boolean;
  setIsUserAdded: (value: boolean) => void;
  canvaUser: UserData | undefined;
  setCanvaUser: (value: UserData) => void;
  isUserInvite: boolean;
  setIsUserInvite: (value: boolean) => void;
  clientRoles: Role[];
  setClientRoles: (value: Role[]) => void;
  isCustomerOnboarding: boolean;
  setIsCustomerOnboarding: (value: boolean) => void;
  zone: Option[];
  setZone: (value: Option[]) => void;
  fetchZones: () => void;
  solinasId: string | undefined;
  setSolinasId: (value: string) => void;
  showSideMenuDesktop: boolean;
  setShowSideMenuDesktop: (value: boolean) => void;
  showOptimizedOffCanvas: boolean;
  setShowOptimizedOffCanvas: (value: boolean) => void;
  optimizedOffCanvasContent: ModalContent | undefined;
  setOptimizedOffCanvasContent: (value: ModalContent) => void;
  showOptimizedPopup: boolean;
  setShowOptimizedPopup: (value: boolean) => void;
  optimizedPopupContent: ModalContent | undefined;
  setOptimizedPopupContent: (value: ModalContent) => void;
  sideMenuRef: HTMLDivElement | null;
  setSideMenuRef: (value: HTMLDivElement) => void;
  activeTimeStamp: string | undefined;
  setActiveTimeStamp: (value: string) => void;
  isSideBarNavigation: boolean;
  setSideBarNavigation: (value: boolean) => void;
  isNavbarNavigation: boolean,
  setIsNavbarNavigation: (value: boolean) => void;
  menuItems: MenuItem[],
  setMenuItems: (value: MenuItem[]) => void;
  adminMenuItems: MenuItem[],
  setAdminMenuItems: (value: MenuItem[]) => void;
  isMenuItemActive:  (value: MenuItem, path: string) => boolean;
  notifications: NotificationState;
  setNotifications: (value: {items: NotificationItem[], unreadCount: number}) => void;
  handleTabSwitch: () => void;
  isViewCommentsClicked: boolean;
  assignedCustomers: AssignedCustomers[] | null;
  setAssignedCustomers: (value: AssignedCustomers[] | null) => void;
  sideBar: {active: number[], tabs: {icon: string; activeIcon: string}[]};
  setSideBar: (value: any) => void;
  handleSideBar: (value: number | number[]) => void;
  handleNewWindow: (value: { type: 'open' | 'close'; id: number }) => void;
  newWindow: {active: number[], clicked: boolean};
  setNewWindow: (value: any) => void;
  showSideBar: boolean;
  setShowSideBar: (value: boolean) => void;
}